Fix type conversion warnings

+++ libspawner/src/options.cpp
@@ -46,7 +46,7 @@ void options_class::add_argument(std::string argument) {
 }
 
 void options_class::add_arguments(const std::vector<std::string> &arguments_a) {
-    for (int i = 0; i < arguments_a.size(); ++i) {
+    for (size_t i = 0; i < arguments_a.size(); ++i) {
         arguments.push_back(arguments_a[i]);
     }
 }
